Replace duration `subsec_nanos` with `as_millis`
the function previously checked whether the elapsed duration between now
and the last timeout check was < 50ms via the `Duration::subsec_nanos`
method. the problem is that `subsec_nanos` is always the fractional
component of the underlying timespec and not the number of elapsed
nanoseconds, which is presumably what was desired. this fix changes the
code to check against the `Duration::as_millis` value, which is the
elapsed number of milliseconds, not fractional. I used `as_millis`
instead of `as_nanos` because `as_nanos` returns `u128`, which seemed
overkill.

finally, it also replaces the `Instant::duration_since` call with
`Instant::saturating_duration_since`, simply because even monotonic
clocks are sometimes not monotonic. `saturating_duration_since` is a
recent addition to the `Instant` api that does not panic if, through
some strange, unexpected error, the time in `self.last_timeout` was
larger than the call to `Instant::now()` in the function.